Solve 2 problems (#34)
* Modifying errors when dealing with opensds.conf

* Automatically set the dashboard configuration file

* Simplify the steps

+++ ansible/script/keystone.sh
@@ -17,28 +17,28 @@
 # 'stack' user is just for install keystone through devstack
 
 create_user(){
-    if id ${STACK_USER_NAME} &> /dev/null; then
+    if id "${STACK_USER_NAME}" &> /dev/null; then
         return
     fi
-    sudo useradd -s /bin/bash -d ${STACK_HOME} -m ${STACK_USER_NAME}
+    sudo useradd -s /bin/bash -d "${STACK_HOME}" -m "${STACK_USER_NAME}"
     echo "stack ALL=(ALL) NOPASSWD: ALL" | sudo tee /etc/sudoers.d/stack
 }
 
 
 remove_user(){
-    userdel ${STACK_USER_NAME} -f -r
+    userdel "${STACK_USER_NAME}" -f -r
     rm /etc/sudoers.d/stack
 }
 
 devstack_local_conf(){
 DEV_STACK_LOCAL_CONF=${DEV_STACK_DIR}/local.conf
-cat > $DEV_STACK_LOCAL_CONF << DEV_STACK_LOCAL_CONF_DOCK
+cat > "$DEV_STACK_LOCAL_CONF" << DEV_STACK_LOCAL_CONF_DOCK
 [[local|localrc]]
 # use TryStack git mirror
 GIT_BASE=$STACK_GIT_BASE
 
-# If the ``*_PASSWORD`` variables are not set here you will be prompted to enter
-# values for them by ``stack.sh``and they will be added to ``local.conf``.
+# If the "*_PASSWORD" variables are not set here you will be prompted to enter
+# values for them by "stack.sh" and they will be added to "local.conf".
 ADMIN_PASSWORD=$STACK_PASSWORD
 DATABASE_PASSWORD=$STACK_PASSWORD
 RABBIT_PASSWORD=$STACK_PASSWORD
@@ -51,7 +51,7 @@ HOST_IP=$HOST_IP
 LOGFILE=\$DEST/logs/stack.sh.log
 
 # Old log files are automatically removed after 7 days to keep things neat.  Change
-# the number of days by setting ``LOGDAYS``.
+# the number of days by setting "LOGDAYS".
 LOGDAYS=2
 
 ENABLED_SERVICES=mysql,key
@@ -60,11 +60,13 @@ ENABLED_SERVICES=mysql,key
 KEYSTONE_BRANCH=$STACK_BRANCH
 KEYSTONECLIENT_BRANCH=$STACK_BRANCH
 DEV_STACK_LOCAL_CONF_DOCK
-chown stack:stack $DEV_STACK_LOCAL_CONF
+chown stack:stack "$DEV_STACK_LOCAL_CONF"
 }
 
 opensds_conf() {
-cat >> $OPENSDS_CONFIG_DIR/opensds.conf << OPENSDS_GLOBAL_CONFIG_DOC
+cat >> "$OPENSDS_CONFIG_DIR/opensds.conf" << OPENSDS_GLOBAL_CONFIG_DOC
+
+
 [keystone_authtoken]
 memcached_servers = $HOST_IP:11211
 signing_dir = /var/cache/opensds
@@ -80,25 +82,25 @@ auth_type = password
 
 OPENSDS_GLOBAL_CONFIG_DOC
 
-cp $OPENSDS_DIR/examples/policy.json $OPENSDS_CONFIG_DIR
+cp "$OPENSDS_DIR/examples/policy.json" "$OPENSDS_CONFIG_DIR"
 }
 
 create_user_and_endpoint(){
-    . $DEV_STACK_DIR/openrc admin admin
-    openstack user create --domain default --password $STACK_PASSWORD $OPENSDS_SERVER_NAME
+    . "$DEV_STACK_DIR/openrc" admin admin
+    openstack user create --domain default --password "$STACK_PASSWORD" "$OPENSDS_SERVER_NAME"
     openstack role add --project service --user opensds admin
     openstack group create service
     openstack group add user service opensds
     openstack role add service --project service --group service
     openstack group add user admins admin
-    openstack service create --name opensds$OPENSDS_VERSION --description "OpenSDS Block Storage" opensds$OPENSDS_VERSION
-    openstack endpoint create --region RegionOne opensds$OPENSDS_VERSION public http://$HOST_IP:50040/$OPENSDS_VERSION/%\(tenant_id\)s
-    openstack endpoint create --region RegionOne opensds$OPENSDS_VERSION internal http://$HOST_IP:50040/$OPENSDS_VERSION/%\(tenant_id\)s
-    openstack endpoint create --region RegionOne opensds$OPENSDS_VERSION admin http://$HOST_IP:50040/$OPENSDS_VERSION/%\(tenant_id\)s
+    openstack service create --name "opensds$OPENSDS_VERSION" --description "OpenSDS Block Storage" "opensds$OPENSDS_VERSION"
+    openstack endpoint create --region RegionOne "opensds$OPENSDS_VERSION" public "http://$HOST_IP:50040/$OPENSDS_VERSION/%\(tenant_id\)s"
+    openstack endpoint create --region RegionOne "opensds$OPENSDS_VERSION" internal "http://$HOST_IP:50040/$OPENSDS_VERSION/%\(tenant_id\)s"
+    openstack endpoint create --region RegionOne "opensds$OPENSDS_VERSION" admin "http://$HOST_IP:50040/$OPENSDS_VERSION/%\(tenant_id\)s"
 }
 
 delete_redundancy_data() {
-    . $DEV_STACK_DIR/openrc admin admin
+    . "$DEV_STACK_DIR/openrc" admin admin
     openstack project delete demo
     openstack project delete alt_demo
     openstack project delete invisible_to_admin
@@ -107,11 +109,10 @@ delete_redundancy_data() {
 }
 
 download_code(){
-    if [ ! -d ${DEV_STACK_DIR} ];then
-        git clone ${STACK_GIT_BASE}/openstack-dev/devstack.git -b ${STACK_BRANCH} ${DEV_STACK_DIR}
-        chown stack:stack -R ${DEV_STACK_DIR}
+    if [ ! -d "${DEV_STACK_DIR}" ];then
+        git clone "${STACK_GIT_BASE}/openstack-dev/devstack.git" -b "${STACK_BRANCH}" "${DEV_STACK_DIR}"
+        chown stack:stack -R "${DEV_STACK_DIR}"
     fi
-
 }
 
 install(){
@@ -120,39 +121,37 @@ install(){
     opensds_conf
 
     # If keystone is ready to start, there is no need continue next step.
-    if wait_for_url http://$HOST_IP/identity "keystone" 0.25 4; then
+    if wait_for_url "http://$HOST_IP/identity" "keystone" 0.25 4; then
         return
     fi
     devstack_local_conf
-    cd ${DEV_STACK_DIR}
-    su $STACK_USER_NAME -c ${DEV_STACK_DIR}/stack.sh
+    cd "${DEV_STACK_DIR}"
+    su "$STACK_USER_NAME" -c "${DEV_STACK_DIR}/stack.sh" >/dev/null
     create_user_and_endpoint
     delete_redundancy_data
 }
 
 cleanup() {
-    su $STACK_USER_NAME -c ${DEV_STACK_DIR}/clean.sh
+    su "$STACK_USER_NAME" -c "${DEV_STACK_DIR}/clean.sh" >/dev/null
 }
 
 uninstall(){
-    su $STACK_USER_NAME -c ${DEV_STACK_DIR}/unstack.sh
+    su "$STACK_USER_NAME" -c "${DEV_STACK_DIR}/unstack.sh" >/dev/null
 }
 
 uninstall_purge(){
-    rm $STACK_HOME/* -rf
+    rm "${STACK_HOME:?'STACK_HOME must be defined and cannot be empty'}/*" -rf
     remove_user
 }
 
 # ***************************
-
-# Keep track of the script directory
 TOP_DIR=$(cd $(dirname "$0") && pwd)
 
 # OpenSDS configuration directory
 OPENSDS_CONFIG_DIR=${OPENSDS_CONFIG_DIR:-/etc/opensds}
 
-source $TOP_DIR/util.sh
-source $TOP_DIR/sdsrc
+source "$TOP_DIR/util.sh"
+source "$TOP_DIR/sdsrc"
 
 case "$# $1" in
     "1 install")
